1.0.0-beta
Doesn't include the gui.
This is a complete rewrite of Autorun. Much nicer codebase, so should be easier to add new features to.

The biggest effort in this update is stability.  There are much less unwraps and unhandled exceptions, and the whole thing is secured to not crash gmod (at least on startup.) If a panic does occur, it will print a proper 'human' error log.

Additionally rglua has been updated from an ancient <1.0 version to 2.1