Basic linux support
Not sure if this works. A *lot* of functionality is stripped out that doesn't work with windows. I can't test it either, I just know the build works with WSL.

* No version checking with HTTP (last commit) because native-tls doesn't seem to work? (need to try again)

* No console ``clear`` command, no ``hide`` or trayicon either.

Binary size is very large, a little short of the size of a debug binary on windows, which is very interesting..